Delete load balancers before removing VPC
If there is any elastic load balancer that holds network interfaces,
it is not possible for NAT/Internet gateways to delete the NICs.
As a result a VPC that depends on the gateways cannot be deleted as
well.

So we should remove every elastic load balancer, before removing VPCs.

As `aws elb load-balancers` does not support `--filters` option,
we should rely on pure jq filters.